refactor(backend): Extracted api from cmd to internal, refactored folder structure
