--- backend/cmd/api/api.go
@@ -1,39 +1,0 @@
-package main
-
-import (
-	"log"
-	"net/http"
-	"time"
-
-	"github.com/FlamingoTP/cardz/internal"
-)
-
-type application struct {
-	config config
-}
-
-type config struct {
-	addr string
-}
-
-func (app *application) mount() *http.ServeMux {
-	mux := http.NewServeMux()
-
-	mux.HandleFunc("GET /api/health", internal.GetHealth)
-
-	return mux
-}
-
-func (app *application) run(mux *http.ServeMux) error {
-	srv := &http.Server{
-		Addr:         app.config.addr,
-		Handler:      mux,
-		WriteTimeout: time.Second * 30,
-		ReadTimeout:  time.Second * 10,
-		IdleTimeout:  time.Minute,
-	}
-
-	log.Printf("Server has started at %s", app.config.addr)
-
-	return srv.ListenAndServe()
-}
