+++ backend/internal/handlers/health.go
@@ -1,4 +1,4 @@
-package internal
+package handlers
 
 import (
 	"encoding/json"
